+++ lib/kitchen/directions/bake_chapter_glossary/v1.rb
@@ -2,8 +2,6 @@
 
 module Kitchen::Directions::BakeChapterGlossary
   class V1
-    renderable
-
     class Definition
       attr_reader :element