+++ lib/kitchen/directions/bake_chapter_key_equations.rb
@@ -15,7 +15,6 @@ module Kitchen
       end
 
       class V1
-        renderable
         def bake(chapter:, metadata_source:, append_to:, uuid_prefix:)
           MoveCustomSectionToEocContainer.v1(
             chapter: chapter,