+++ lib/kitchen/directions/bake_chapter_solutions/v1.rb
@@ -1,16 +1,8 @@
 # frozen_string_literal: true
 
 module Kitchen::Directions::BakeChapterSolutions
-  # TODO: Refactor to use EocCompositePageContainer or MoveCustomSectionToEocContainer
   class V1
-    renderable
-
     def bake(chapter:, metadata_source:, uuid_prefix: '')
-      @metadata = metadata_source.children_to_keep.copy
-      @klass = 'solutions'
-      @title = I18n.t(:eoc_solutions_title)
-      @uuid_prefix = uuid_prefix
-
       solutions_clipboard = Kitchen::Clipboard.new
 
       chapter.search('section.free-response').each do |free_response_question|
@@ -27,12 +19,15 @@ module Kitchen::Directions::BakeChapterSolutions
         end
       end
 
-      @content = solutions_clipboard.paste
+      content = solutions_clipboard.paste
 
-      @in_composite_chapter = false
-
-      chapter.append(child: render(file:
-        '../../templates/eoc_section_template_old.xhtml.erb'))
+      Kitchen::Directions::EocCompositePageContainer.v1(
+        container_key: 'solutions',
+        uuid_key: "#{uuid_prefix}solutions",
+        metadata_source: metadata_source,
+        content: content,
+        append_to: chapter
+      )
     end
   end
 end