+++ lib/kitchen/directions/move_exercises_to_eoc/v1.rb
@@ -1,37 +1,19 @@
 # frozen_string_literal: true
 
 module Kitchen::Directions::MoveExercisesToEOC
-  # TODO: Refactor to use EocCompositePageContainer or MoveCustomSectionToEocContainer
   class V1
-    renderable
-
     def bake(chapter:, metadata_source:, klass:, append_to: nil, uuid_prefix: '.')
-      @klass = klass
-      @metadata = metadata_source.children_to_keep.copy
-      @title = I18n.t(:"eoc.#{klass}")
-      @uuid_prefix = uuid_prefix
-
-      exercise_clipboard = Kitchen::Clipboard.new
-
-      chapter.non_introduction_pages.each do |page|
-        sections = page.search("section.#{@klass}")
-
-        sections.each do |exercise_section|
-          exercise_section.first("[data-type='title']")&.trash
-
-          exercise_section.cut(to: exercise_clipboard)
-        end
+      Kitchen::Directions::MoveCustomSectionToEocContainer.v1(
+        chapter: chapter,
+        metadata_source: metadata_source,
+        container_key: klass,
+        uuid_key: "#{uuid_prefix}#{klass}",
+        section_selector: "section.#{klass}",
+        append_to: append_to || chapter,
+        include_intro_page: false
+      ) do |exercise_section|
+        Kitchen::Directions::RemoveSectionTitle.v1(section: exercise_section)
       end
-
-      return if exercise_clipboard.none?
-
-      @content = exercise_clipboard.paste
-
-      append_to_element = append_to || chapter
-      @in_composite_chapter = append_to_element.is?(:composite_chapter)
-
-      append_to_element.append(child: render(file:
-        '../../templates/eoc_section_template_old.xhtml.erb'))
     end
   end
 end